Add strings.xml file for project's strings
This commit adds the `strings.xml` file to the project, which contains all the string resources used throughout the app. The `strings.xml` file includes the strings used in the login screen, registration screen, and home screen of the app.

This change is a logical unit of work that will make it easier to manage and update the strings as the project evolves.