Add colors.xml file for project's colors and modify themes.xml file for update colors.
This commit adds the `colors.xml` file to the project, which contains all the colors used throughout the app. The `colors.xml` file includes the colors used in the hole app.

This change is a logical unit of work that will make it easier to manage and use the  colors as the project evolves.